requirements.txt
# ===================
# API Framework
# ===================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# ===================
# ML Model
# ===================
transformers>=4.36.0
torch>=2.0.0
scipy>=1.11.0
numpy>=1.24.0
scikit-learn>=1.3.0

# ===================
# Agent Framework
# ===================
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# ===================
# Chat UI
# ===================
chainlit>=1.0.0

# ===================
# HTTP Client
# ===================
httpx>=0.26.0
requests>=2.31.0

# ===================
# Data Processing
# ===================
pandas>=2.0.0

# ===================
# Testing
# ===================
pytest>=7.4.0
pytest-asyncio>=0.23.0

# ===================
# Optional: MCP Support
# ===================
# fastapi-mcp>=0.2.0
